refactor(client): use query param over header for client version reporting
