+++ oprf-client/src/ws/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(not(target_arch = "wasm32"))]
 mod native;
+use http::Uri;
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) use native::WebSocketSession;
 
@@ -7,3 +8,14 @@ pub(crate) use native::WebSocketSession;
 mod wasm;
 #[cfg(target_arch = "wasm32")]
 pub(crate) use wasm::WebSocketSession;
+
+pub(crate) fn append_client_version_to_query(endpoint: &Uri) -> String {
+    let version = env!("CARGO_PKG_VERSION");
+    let has_query = endpoint.query().is_some();
+    let mut endpoint = endpoint.to_string();
+
+    endpoint.push(if has_query { '&' } else { '?' });
+    endpoint.push_str("version=");
+    endpoint.push_str(version);
+    endpoint
+}
